+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-05-06
+
+### Added
+- New config Mix_ExcludeRegrowing, prevent mixed seeds from planting regrowing crops.
+- New config Mix_ExcludeRare, prevent mixed seeds from planting seeds you can't buy from Pierre's/Joja.
+
 ## [0.1.0] - 2026-04-18
 
 ### Added